requirements.txt
walkdir
requests
beautifulsoup4[lxml]
mysql-connector-python